Move CarteEtat to TS
